feat(tree-node-web): create TreeNodeBranch component
