+++ packages/pluggableWidgets/tree-node-web/src/components/TreeNode.tsx
@@ -1,36 +1,14 @@
-import { WebIcon, ObjectItem } from "mendix";
-import {
-    createElement,
-    CSSProperties,
-    HTMLAttributes,
-    ReactElement,
-    ReactEventHandler,
-    ReactNode,
-    SyntheticEvent,
-    useCallback,
-    useContext,
-    useEffect,
-    useLayoutEffect,
-    useRef,
-    useState
-} from "react";
 import classNames from "classnames";
-import { OpenNodeOnEnum, ShowIconEnum, TreeNodeContainerProps } from "../../typings/TreeNodeProps";
-import {
-    TreeNodeBranchContextProps,
-    TreeNodeBranchContext,
-    useInformParentContextOfChildNodes
-} from "./TreeNodeBranchContext";
+import { WebIcon, ObjectItem } from "mendix";
+import { createElement, CSSProperties, ReactElement, ReactNode, useCallback, useContext } from "react";
 
-import {
-    TreeNodeFocusChangeHandler,
-    useTreeNodeBranchKeyboardHandler,
-    useTreeNodeFocusChangeHandler
-} from "./hooks/TreeNodeAccessibility";
-import { useTreeNodeLazyLoading } from "./hooks/lazyLoading";
-import { useAnimatedTreeNodeContentHeight } from "./hooks/useAnimatedHeight";
+import { OpenNodeOnEnum, TreeNodeContainerProps } from "../../typings/TreeNodeProps";
+
+import { useTreeNodeFocusChangeHandler } from "./hooks/TreeNodeAccessibility";
 import { useTreeNodeRef } from "./hooks/useTreeNodeRef";
 import { renderTreeNodeHeaderIcon, TreeNodeHeaderIcon } from "./HeaderIcon";
+import { TreeNodeBranch, TreeNodeBranchProps, treeNodeBranchUtils } from "./TreeNodeBranch";
+import { TreeNodeBranchContext, useInformParentContextOfChildNodes } from "./TreeNodeBranchContext";
 
 export interface TreeNodeItem extends ObjectItem {
     headerContent: ReactNode;
@@ -120,192 +98,10 @@ export function TreeNode({
         </ul>
     );
 }
-interface TreeNodeBranchProps {
-    id: TreeNodeItem["id"];
-    isUserDefinedLeafNode: boolean;
-    startExpanded: boolean;
-    headerContent: ReactNode;
-    openNodeOn: OpenNodeOnEnum;
-    children: ReactNode;
-    iconPlacement: ShowIconEnum;
-    renderHeaderIcon: (treeNodeState: TreeNodeState, iconPlacement: Exclude<ShowIconEnum, "no">) => ReactNode;
-    changeFocus: TreeNodeFocusChangeHandler;
-    animateTreeNodeContent: boolean;
-}
-
-const treeNodeBranchUtils = {
-    bodyClassName: "widget-tree-node-body",
-    getHeaderId: (id: TreeNodeItem["id"]) => `${id}TreeNodeBranchHeader`,
-    getBodyId: (id: TreeNodeItem["id"]) => `${id}TreeNodeBranchBody`
-};
-
-function getTreeNodeAccessibilityProps(isExpanded: boolean): HTMLAttributes<HTMLLIElement> {
-    return {
-        "aria-expanded": isExpanded,
-        role: "treeitem",
-        tabIndex: 0
-    };
-}
 
 export const enum TreeNodeState {
     COLLAPSED_WITH_JS = "COLLAPSED_WITH_JS",
     COLLAPSED_WITH_CSS = "COLLAPSED_WITH_CSS",
     EXPANDED = "EXPANDED",
     LOADING = "LOADING"
-}
-
-function TreeNodeBranch(props: TreeNodeBranchProps): ReactElement {
-    const { level: currentContextLevel } = useContext(TreeNodeBranchContext);
-    const [treeNodeState, setTreeNodeState] = useState<TreeNodeState>(
-        props.startExpanded ? TreeNodeState.EXPANDED : TreeNodeState.COLLAPSED_WITH_JS
-    );
-    const [isActualLeafNode, setIsActualLeafNode] = useState<boolean>(props.isUserDefinedLeafNode || !props.children);
-    const isHeaderClickable = props.openNodeOn === "headerClick";
-    const isIconClickable = props.openNodeOn === "iconClick";
-
-    useEffect(() => {
-        setIsActualLeafNode(props.isUserDefinedLeafNode || !props.children);
-    }, [props.children]);
-
-    const informParentOfChildNodes = useCallback<TreeNodeBranchContextProps["informParentOfChildNodes"]>(
-        numberOfNodes => {
-            if (numberOfNodes !== undefined) {
-                setTreeNodeState(treeNodeState =>
-                    treeNodeState === TreeNodeState.LOADING ? TreeNodeState.EXPANDED : treeNodeState
-                );
-                setIsActualLeafNode(currentIsActualLeafNode => {
-                    if (numberOfNodes === 0 && !currentIsActualLeafNode) {
-                        return true;
-                    } else if (numberOfNodes > 0 && currentIsActualLeafNode) {
-                        return false;
-                    }
-                    return currentIsActualLeafNode;
-                });
-            }
-        },
-        []
-    );
-
-    const treeNodeBranchBody = useRef<HTMLDivElement>(null);
-    const treeNodeBranchRef = useRef<HTMLLIElement>(null);
-
-    const { hasNestedTreeNode } = useTreeNodeLazyLoading(treeNodeBranchBody);
-    const { isAnimating, captureElementHeight, animateTreeNodeContent, cleanupAnimation } =
-        useAnimatedTreeNodeContentHeight(treeNodeBranchBody);
-
-    useLayoutEffect(() => {
-        if (props.animateTreeNodeContent && treeNodeState !== TreeNodeState.LOADING) {
-            const animationCleanup = animateTreeNodeContent();
-            if (animationCleanup) {
-                return animationCleanup;
-            }
-        }
-    }, [animateTreeNodeContent, props.animateTreeNodeContent, treeNodeState]);
-
-    const eventTargetIsNotCurrentBranch = useCallback<(event: SyntheticEvent<HTMLElement>) => boolean>(event => {
-        const target = event.target as Node;
-        return (
-            !treeNodeBranchRef.current?.isSameNode(target) &&
-            !treeNodeBranchRef.current?.firstElementChild?.contains(target) &&
-            !treeNodeBranchRef.current?.lastElementChild?.isSameNode(target)
-        );
-    }, []);
-
-    useEffect(() => {
-        if (treeNodeState === TreeNodeState.LOADING) {
-            if (!hasNestedTreeNode()) {
-                setTreeNodeState(TreeNodeState.EXPANDED);
-            }
-        }
-    }, [hasNestedTreeNode, treeNodeState]);
-
-    const toggleTreeNodeContent = useCallback<ReactEventHandler<HTMLElement>>(
-        event => {
-            if (eventTargetIsNotCurrentBranch(event)) {
-                return;
-            }
-            if (!isActualLeafNode) {
-                captureElementHeight();
-                setTreeNodeState(treeNodeState => {
-                    if (treeNodeState === TreeNodeState.LOADING) {
-                        // TODO:
-                        return treeNodeState;
-                    }
-                    if (treeNodeState === TreeNodeState.COLLAPSED_WITH_JS) {
-                        return TreeNodeState.LOADING;
-                    }
-                    if (treeNodeState === TreeNodeState.COLLAPSED_WITH_CSS) {
-                        return TreeNodeState.EXPANDED;
-                    }
-                    return TreeNodeState.COLLAPSED_WITH_CSS;
-                });
-            }
-        },
-        [isActualLeafNode, eventTargetIsNotCurrentBranch, captureElementHeight]
-    );
-
-    const treeNodeAccessibilityProps = getTreeNodeAccessibilityProps(treeNodeState === TreeNodeState.EXPANDED);
-
-    const onHeaderKeyDown = useTreeNodeBranchKeyboardHandler(
-        toggleTreeNodeContent,
-        props.changeFocus,
-        treeNodeState,
-        isActualLeafNode,
-        eventTargetIsNotCurrentBranch
-    );
-
-    const onIconClick = isIconClickable ? toggleTreeNodeContent : undefined;
-    const onHeaderClick = isHeaderClickable ? toggleTreeNodeContent : undefined;
-
-    return (
-        <li
-            className="widget-tree-node-branch"
-            onKeyDown={onHeaderKeyDown}
-            ref={treeNodeBranchRef}
-            {...treeNodeAccessibilityProps}
-        >
-            <span
-                className={classNames("widget-tree-node-branch-header", {
-                    "widget-tree-node-branch-header-clickable": !isActualLeafNode && isHeaderClickable,
-                    "widget-tree-node-branch-header-reversed": props.iconPlacement === "left"
-                })}
-                id={treeNodeBranchUtils.getHeaderId(props.id)}
-                onClick={onHeaderClick}
-            >
-                <span className="widget-tree-node-branch-header-value">{props.headerContent}</span>
-                {!isActualLeafNode && props.iconPlacement !== "no" && (
-                    <span
-                        className={classNames("widget-tree-node-branch-header-icon-container", {
-                            "widget-tree-node-branch-header-clickable": !isActualLeafNode && isIconClickable
-                        })}
-                        onClick={onIconClick}
-                    >
-                        {props.renderHeaderIcon(treeNodeState, props.iconPlacement)}
-                    </span>
-                )}
-            </span>
-            {((!isActualLeafNode && treeNodeState !== TreeNodeState.COLLAPSED_WITH_JS) || isAnimating) && (
-                <TreeNodeBranchContext.Provider
-                    value={{
-                        level: currentContextLevel + 1,
-                        informParentOfChildNodes
-                    }}
-                >
-                    <div
-                        className={classNames(treeNodeBranchUtils.bodyClassName, {
-                            "widget-tree-node-branch-hidden":
-                                treeNodeState === TreeNodeState.COLLAPSED_WITH_CSS && !isAnimating,
-                            "widget-tree-node-branch-loading": treeNodeState === TreeNodeState.LOADING
-                        })}
-                        id={treeNodeBranchUtils.getBodyId(props.id)}
-                        aria-hidden={treeNodeState !== TreeNodeState.EXPANDED}
-                        ref={treeNodeBranchBody}
-                        onTransitionEnd={cleanupAnimation}
-                    >
-                        {props.children}
-                    </div>
-                </TreeNodeBranchContext.Provider>
-            )}
-        </li>
-    );
 }
